fixes + tweaks
client fixes:
- deagle laser displayed while deploying
- deagle laser not always disabled when switching weapons
- deagle/rpg laser broken with custom models
- rpg laser state desyncing with the server
- thirdperson camera spinning at light speed with m_rawinput 1
- thirdperson camera not respecting mouse inversion setting
- current/next map scoreboard text overlapping server name at some resolutions
- integer overflow in team scores
- uzi deploying with 3 hands when switched to from knife

server fixes:
- deagle not precached if the map has an otis but not weapon_eagle
- possible fix for suit HUD not displaying
- unusable 556 ammo spawning on HL players
- minigun pickable while empty but can't be selected

server tweaks:
- lighter minigun distant sound
- uzi volume reduced to 60% from 70%

+++ cl_dll/hud.h
@@ -267,8 +267,8 @@ struct extra_player_info_t
 struct team_info_t 
 {
 	char name[MAX_TEAM_NAME];
-	short frags;
-	short deaths;
+	int frags;
+	int deaths;
 	short ping;
 	short packetloss;
 	short ownteam;
