+++ cl_dll/vgui_ScorePanel.h
@@ -155,7 +155,7 @@ public:
 		strcpy(buf, text);
 
 		int len = strlen(buf);
-		while (len && isspace(buf[--len]))
+		while (len && isspace((uint8_t)buf[--len]))
 		{
 			buf[len] = 0;
 		}
